requirements.txt
Flask==2.2.5
Flask-SocketIO==5.3.3
python-socketio==5.10.0
eventlet==0.33.0
flask-sqlalchemy
python-dotenv
dnspython==1.16.0